% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{is_sysreqr_plan}
\alias{is_sysreqr_plan}
\title{Test whether an object is a sysreqr plan}
\usage{
is_sysreqr_plan(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if \code{x} inherits from \code{"sysreqr_plan"}.
}
\description{
Test whether an object is a sysreqr plan
}
\examples{
plan <- check_packages("xml2", platform = "ubuntu-22.04")
is_sysreqr_plan(plan)
is_sysreqr_plan(data.frame(x = 1))
}
\seealso{
Other plan:
\code{\link[=as_data_frame]{as_data_frame()}},
\code{\link[=print.sysreqr_plan]{print.sysreqr_plan()}}
}
\concept{plan}
