% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{print.sysreqr_plan}
\alias{print.sysreqr_plan}
\title{Print a sysreqr plan}
\usage{
\method{print}{sysreqr_plan}(x, ...)
}
\arguments{
\item{x}{A \code{sysreqr_plan}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a sysreqr plan
}
\seealso{
Other plan:
\code{\link[=as_data_frame]{as_data_frame()}},
\code{\link[=is_sysreqr_plan]{is_sysreqr_plan()}}
}
\concept{plan}
