% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-advice.R
\name{print.sysreqr_setup_advice}
\alias{print.sysreqr_setup_advice}
\title{Print setup advice}
\usage{
\method{print}{sysreqr_setup_advice}(x, ...)
}
\arguments{
\item{x}{A \code{sysreqr_setup_advice} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print setup advice
}
\seealso{
Other setup:
\code{\link[=explain]{explain()}},
\code{\link[=setup_advice]{setup_advice()}}
}
\concept{setup}
