% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_dockerfile_snippet}
\alias{write_dockerfile_snippet}
\title{Write a Dockerfile snippet}
\usage{
write_dockerfile_snippet(plan, path)
}
\arguments{
\item{plan}{A \code{sysreqr_plan}.}

\item{path}{Output path.}
}
\value{
\code{path}, invisibly.
}
\description{
Writes the output of \code{\link[=dockerfile]{dockerfile()}} to a file so it can be appended to an
existing Dockerfile or included verbatim.
}
\examples{
plan <- check_packages("xml2", platform = "ubuntu-22.04")
write_dockerfile_snippet(plan, file.path(tempdir(), "Dockerfile.sysreqs"))
}
\seealso{
Other output:
\code{\link[=as_install_plan]{as_install_plan()}},
\code{\link[=write_install_script]{write_install_script()}},
\code{\link[=write_json]{write_json()}},
\code{\link[=write_report]{write_report()}}
}
\concept{output}
