% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{ppm_platforms}
\alias{ppm_platforms}
\title{List Posit Package Manager platforms}
\usage{
ppm_platforms(base_url = ppm_default_base_url())
}
\arguments{
\item{base_url}{Posit Package Manager base URL.}
}
\value{
A data frame of platform records reported by Package Manager.
}
\description{
Queries the Posit Package Manager \verb{/__api__/status} endpoint and returns a
data frame of supported distributions.
}
\examples{
\donttest{
ppm_platforms()
}
}
\seealso{
Other ppm:
\code{\link[=check_ppm]{check_ppm()}},
\code{\link[=ppm_repo]{ppm_repo()}},
\code{\link[=ppm_sysreqs]{ppm_sysreqs()}},
\code{\link[=use_ppm]{use_ppm()}}
}
\concept{ppm}
